requirements.txt
Flask
trimesh
pyrender
numpy
pillow
scipy
pygltflib
